requirements.txt
PyQt5==5.15.9
PyQt5-sip==12.13.0
requests==2.31.0
dnspython==2.4.2
nmap==0.0.1
python-nmap==0.7.1
scapy==2.5.0
geoip2==4.7.0
maxminddb==2.2.0
cryptography==41.0.7
paramiko==3.4.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
openpyxl==3.11.0
sqlalchemy==2.0.23
pillow==10.1.0
psutil==5.9.6
whois==0.9.7
reportlab==4.0.7
jinja2==3.1.2
requests-html==0.10.0
packaging==23.2
